Move listener adapter to dedicated file
