+++ dds/src/implementation/utils/mod.rs
@@ -1,2 +1,3 @@
 pub mod actor;
 pub mod instance_handle_from_key;
+pub mod sync_listener;
